(PC-40074)[PRO] refactor: reorder files
