+++ pro/src/pages/Homepage/components/StatsCard/__specs__/CumulatedViews.spec.tsx
