+++ pro/src/pages/Homepage/components/StatsCard/__specs__/MostViewedOffers.spec.tsx
