+++ pro/src/pages/Homepage/components/StatsCard/components/CumulatedViews.module.scss
