+++ pro/src/pages/Homepage/components/StatsCard/components/CumulatedViewsEmptyState.tsx
