+++ pro/src/pages/Offerers/Offerer/VenueV1/VenueEdition/CollectiveDataEdition/CollectiveDataEdition.tsx
@@ -8,7 +8,7 @@ import { useAppSelector } from '@/commons/hooks/useAppSelector'
 import { ensureSelectedPartnerVenue } from '@/commons/store/user/selectors'
 import { CollectiveDmsTimeline } from '@/components/CollectiveDmsTimeline/CollectiveDmsTimeline'
 import { Banner } from '@/design-system/Banner/Banner'
-import { PartnerPageCollectiveSection } from '@/pages/Homepage/components/Offerers/components/PartnerPages/components/PartnerPageCollectiveSection'
+import { PartnerPageCollectiveSection } from '@/pages/Offerers/Offerer/VenueV1/VenueEdition/CollectiveDataEdition/PartnerPageCollectiveSection/PartnerPageCollectiveSection'
 import { Spinner } from '@/ui-kit/Spinner/Spinner'
 
 import styles from './CollectiveDataEdition.module.scss'
