+++ pro/src/pages/Offerers/Offerer/VenueV1/VenueEdition/CollectiveDataEdition/PartnerPageCollectiveSection/PartnerPageCollectiveSection.module.scss
@@ -2,39 +2,6 @@
 @use "styles/mixins/_rem.scss" as rem;
 @use "styles/mixins/_size.scss" as size;
 
-.header {
-  display: flex;
-  flex-direction: column;
-  gap: var(--size-spacing-xl);
-  margin-bottom: var(--size-spacing-xxl);
-
-  @media (min-width: size.$tablet) {
-    flex-direction: row;
-    align-items: center;
-  }
-}
-
-.image-uploader {
-  align-self: center;
-  margin-bottom: 0;
-}
-
-.venue {
-  height: fit-content;
-  gap: var(--size-spacing-s);
-  display: flex;
-  flex-direction: column;
-
-  &-type {
-    color: var(--color-text-subtle);
-    text-transform: uppercase;
-  }
-
-  &-name {
-    @include fonts.title3;
-  }
-}
-
 .details {
   display: flex;
   flex-direction: column;
